backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==2.3.0
pydantic[email]>=2.0.0
gunicorn==21.2.0
requests>=2.31.0
PyJWT>=2.0.0
httpx>=0.25
httpcore>=0.18
bcrypt==4.1.2
numpy>=1.24.0
